requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
folium>=0.15.0
streamlit-folium>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
